requirements.txt
python-telegram-bot[rate-limiter]==20.7
